+++ changelog.md
@@ -10,4 +10,4 @@
 ## 0.5
 - Switches - power outlet, digital out
 ## 0.6
-- Sensors - Humidity, Temperature, Variable, Motion
+- Sensors - Humidity, Temperature, Variable, Motion, Generic reed contact, Rain counter, Electric counter (consumption, demand), Brightness, Co2, Wind speed, Pulse counter (total impulse, current hour impulse, frequency)